+++ src/contracts/IChain.sol
@@ -1,56 +1,63 @@
 // SPDX-License-Identifier: MIT
-pragma solidity >=0.6.8 <=0.8.0;
+pragma solidity >=0.6.8;
 pragma abicoder v2;
 
 interface IChain {
   struct Block {
     bytes32 root;
-    address minter;
-    uint256 staked;
-    uint256 power;
-    uint256 anchor;
-    uint256 timestamp;
-    uint256 dataTimestamp;
+    uint32 dataTimestamp;
+    uint128 affidavit;
   }
 
-  function blockPadding() external view returns (uint256);
+  struct FirstClassData {
+    uint224 value;
+    uint32 dataTimestamp;
+  }
 
-  function blocksCount() external view returns (uint256);
+  function blocks(uint256) external view returns (Block memory);
+  function fcds(bytes32) external view returns (FirstClassData memory);
+
+  function blocksCount() external view returns (uint32);
+
+  function blocksCountOffset() external view returns (uint32);
+
+  function padding() external view returns (uint16);
 
   function getName() external pure returns (bytes32);
 
   function recoverSigner(bytes32 affidavit, uint8 _v, bytes32 _r, bytes32 _s) external pure returns (address);
 
-  function getStatus() external view returns (
+  function getStatus() external view returns(
     uint256 blockNumber,
-    uint256 lastDataTimestamp,
-    uint256 lastBlockHeight,
+    uint16 timePadding,
+    uint32 lastDataTimestamp,
+    uint32 lastBlockId,
     address nextLeader,
-    uint256 nextBlockHeight,
+    uint32 nextBlockId,
     address[] memory validators,
     uint256[] memory powers,
     string[] memory locations,
     uint256 staked
   );
 
-  function getBlockHeight() external view returns (uint256);
+  function getBlockId() external view returns (uint32);
 
-  function getLatestBlockHeightWithData() external view returns (uint256);
+  function getLatestBlockId() external view returns (uint32);
 
-  function getLeaderIndex(uint256 numberOfValidators, uint256 ethBlockNumber) external view returns (uint256);
+  function getLeaderIndex(uint256 _numberOfValidators, uint256 _timestamp) external view returns (uint256);
 
   function getNextLeaderAddress() external view returns (address);
-  
-  function getLeaderAddress() external view returns (address);
-  
-  function getLeaderAddressAtBlock(uint256 ethBlockNumber) external view returns (address);
 
-  function verifyProof(bytes32[] calldata _proof, bytes32 _root, bytes32 _leaf) external pure returns (bool);
+  function getLeaderAddress() external view returns (address);
+
+  function getLeaderAddressAtTime(uint232 _timestamp) external view returns (address);
 
   function hashLeaf(bytes calldata _key, bytes calldata _value) external pure returns (bytes32);
 
+  function verifyProof(bytes32[] calldata _proof, bytes32 _root, bytes32 _leaf) external pure returns (bool);
+
   function verifyProofForBlock(
-    uint256 _blockHeight,
+    uint256 _blockId,
     bytes32[] calldata _proof,
     bytes calldata _key,
     bytes calldata _value
@@ -62,56 +69,12 @@ interface IChain {
     uint256 _items
   ) external pure returns (bytes32[] memory);
 
-  function verifyProofs(
-    uint256[] calldata _blockHeights,
-    bytes calldata _proofs,
-    uint256[] calldata _proofItemsCounter,
-    bytes32[] calldata _leaves
-  ) external view returns (bool[] memory results);
-  
-  function decodeLeafToNumber(bytes calldata _leaf) external pure returns (uint);
+  function getBlockRoot(uint256 _blockId) external view returns (bytes32);
 
-  function decodeLeafToFloat(bytes calldata _leaf) external pure returns (uint);
+  function getBlockTimestamp(uint32 _blockId) external view returns (uint32);
 
-  function verifyProofForBlockForNumber(
-    uint256 _blockHeight,
-    bytes32[] calldata _proof,
-    bytes calldata _key,
-    bytes calldata _value
-  ) external returns (bool, uint256);
+  function getCurrentValues(bytes32[] calldata _keys)
+  external view returns (uint256[] memory values, uint256[] memory timestamps);
 
-  function verifyProofForBlockForFloat(
-    uint256 _blockHeight,
-    bytes32[] calldata _proof,
-    bytes calldata _key,
-    bytes calldata _value
-  ) external view returns (bool, uint256);
-
-  function getBlockData(uint256 _blockHeight) external view returns (Block memory);
-
-  function getBlockRoot(uint256 _blockHeight) external view returns (bytes32);
-
-  function getBlockMinter(uint256 _blockHeight) external view returns (address);
-
-  function getBlockStaked(uint256 _blockHeight) external view returns (uint256);
-
-  function getBlockPower(uint256 _blockHeight) external view returns (uint256);
-
-  function getBlockAnchor(uint256 _blockHeight) external view returns (uint256);
-
-  function getBlockTimestamp(uint256 _blockHeight) external view returns (uint256);
-
-  function getBlockVotersCount(uint256 _blockHeight) external view returns (uint256);
-
-  function getBlockVoters(uint256 _blockHeight) external view returns (address[] memory);
-
-  function getBlockVotes(uint256 _blockHeight, address _voter) external view returns (uint256);
-
-  function getNumericFCD(uint256 _blockHeight, bytes32 _key) external view returns (uint256 value, uint timestamp);
-
-  function getNumericFCDs(
-    uint256 _blockHeight, bytes32[] calldata _keys
-  ) external view returns (uint256[] memory values, uint256 timestamp);
-
-  function getCurrentValue(bytes32 _key) external view returns (uint256 value, uint timestamp);
+  function getCurrentValue(bytes32 _key) external view returns (uint256 value, uint256 timestamp);
 }