+++ src/contracts/abi/registry.abi.ts
@@ -1,199 +1,199 @@
 export const registryAbi = [
   {
-    'anonymous': false,
-    'inputs': [
+    anonymous: false,
+    inputs: [
       {
-        'indexed': true,
-        'internalType': 'address',
-        'name': 'destination',
-        'type': 'address'
+        indexed: true,
+        internalType: 'address',
+        name: 'destination',
+        type: 'address',
       },
       {
-        'indexed': false,
-        'internalType': 'bytes32',
-        'name': 'name',
-        'type': 'bytes32'
-      }
+        indexed: false,
+        internalType: 'bytes32',
+        name: 'name',
+        type: 'bytes32',
+      },
     ],
-    'name': 'LogRegistered',
-    'type': 'event'
+    name: 'LogRegistered',
+    type: 'event',
   },
   {
-    'anonymous': false,
-    'inputs': [
+    anonymous: false,
+    inputs: [
       {
-        'indexed': true,
-        'internalType': 'address',
-        'name': 'previousOwner',
-        'type': 'address'
+        indexed: true,
+        internalType: 'address',
+        name: 'previousOwner',
+        type: 'address',
       },
       {
-        'indexed': true,
-        'internalType': 'address',
-        'name': 'newOwner',
-        'type': 'address'
-      }
+        indexed: true,
+        internalType: 'address',
+        name: 'newOwner',
+        type: 'address',
+      },
     ],
-    'name': 'OwnershipTransferred',
-    'type': 'event'
+    name: 'OwnershipTransferred',
+    type: 'event',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'bytes32',
-        'name': '_bytes',
-        'type': 'bytes32'
-      }
+        internalType: 'bytes32',
+        name: '_bytes',
+        type: 'bytes32',
+      },
     ],
-    'name': 'getAddress',
-    'outputs': [
+    name: 'getAddress',
+    outputs: [
       {
-        'internalType': 'address',
-        'name': '',
-        'type': 'address'
-      }
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
     ],
-    'stateMutability': 'view',
-    'type': 'function'
+    stateMutability: 'view',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'string',
-        'name': '_name',
-        'type': 'string'
-      }
+        internalType: 'string',
+        name: '_name',
+        type: 'string',
+      },
     ],
-    'name': 'getAddressByString',
-    'outputs': [
+    name: 'getAddressByString',
+    outputs: [
       {
-        'internalType': 'address',
-        'name': '',
-        'type': 'address'
-      }
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
     ],
-    'stateMutability': 'view',
-    'type': 'function'
+    stateMutability: 'view',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'bytes32[]',
-        'name': '_names',
-        'type': 'bytes32[]'
+        internalType: 'bytes32[]',
+        name: '_names',
+        type: 'bytes32[]',
       },
       {
-        'internalType': 'address[]',
-        'name': '_destinations',
-        'type': 'address[]'
-      }
+        internalType: 'address[]',
+        name: '_destinations',
+        type: 'address[]',
+      },
     ],
-    'name': 'importAddresses',
-    'outputs': [],
-    'stateMutability': 'nonpayable',
-    'type': 'function'
+    name: 'importAddresses',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'address[]',
-        'name': '_destinations',
-        'type': 'address[]'
-      }
+        internalType: 'address[]',
+        name: '_destinations',
+        type: 'address[]',
+      },
     ],
-    'name': 'importContracts',
-    'outputs': [],
-    'stateMutability': 'nonpayable',
-    'type': 'function'
+    name: 'importContracts',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
   },
   {
-    'inputs': [],
-    'name': 'owner',
-    'outputs': [
+    inputs: [],
+    name: 'owner',
+    outputs: [
       {
-        'internalType': 'address',
-        'name': '',
-        'type': 'address'
-      }
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
     ],
-    'stateMutability': 'view',
-    'type': 'function'
+    stateMutability: 'view',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'bytes32',
-        'name': '',
-        'type': 'bytes32'
-      }
+        internalType: 'bytes32',
+        name: '',
+        type: 'bytes32',
+      },
     ],
-    'name': 'registry',
-    'outputs': [
+    name: 'registry',
+    outputs: [
       {
-        'internalType': 'address',
-        'name': '',
-        'type': 'address'
-      }
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
     ],
-    'stateMutability': 'view',
-    'type': 'function'
+    stateMutability: 'view',
+    type: 'function',
   },
   {
-    'inputs': [],
-    'name': 'renounceOwnership',
-    'outputs': [],
-    'stateMutability': 'nonpayable',
-    'type': 'function'
+    inputs: [],
+    name: 'renounceOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'bytes32',
-        'name': 'name',
-        'type': 'bytes32'
-      }
+        internalType: 'bytes32',
+        name: 'name',
+        type: 'bytes32',
+      },
     ],
-    'name': 'requireAndGetAddress',
-    'outputs': [
+    name: 'requireAndGetAddress',
+    outputs: [
       {
-        'internalType': 'address',
-        'name': '',
-        'type': 'address'
-      }
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
     ],
-    'stateMutability': 'view',
-    'type': 'function'
+    stateMutability: 'view',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'string',
-        'name': '_string',
-        'type': 'string'
-      }
+        internalType: 'string',
+        name: '_string',
+        type: 'string',
+      },
     ],
-    'name': 'stringToBytes32',
-    'outputs': [
+    name: 'stringToBytes32',
+    outputs: [
       {
-        'internalType': 'bytes32',
-        'name': 'result',
-        'type': 'bytes32'
-      }
+        internalType: 'bytes32',
+        name: 'result',
+        type: 'bytes32',
+      },
     ],
-    'stateMutability': 'pure',
-    'type': 'function'
+    stateMutability: 'pure',
+    type: 'function',
   },
   {
-    'inputs': [
+    inputs: [
       {
-        'internalType': 'address',
-        'name': 'newOwner',
-        'type': 'address'
-      }
+        internalType: 'address',
+        name: 'newOwner',
+        type: 'address',
+      },
     ],
-    'name': 'transferOwnership',
-    'outputs': [],
-    'stateMutability': 'nonpayable',
-    'type': 'function'
-  }
+    name: 'transferOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
 ];