+++ src/converters/index.ts
@@ -1,2 +1,0 @@
-export {strToBytes32} from './strToBytes32';
-export {numberToUint256} from './numberToUint256';