+++ src/index.ts
@@ -1,21 +1,19 @@
-import {LeafType} from './models/LeafType';
-import {IAPIClientOptions} from './models/APIClientOptions';
-import {LeafKeyCoder} from './services/LeafKeyCoder';
-import {APIClient} from './services/APIClient';
-import {LeafValueCoder} from './services/LeafValueCoder';
-import {Registry as ContractRegistry} from './contracts/Registry';
-import {ChainContract} from './contracts/Chain';
-import * as converters from './converters';
+import { IAPIClientOptions } from './models/APIClientOptions';
+import { LeafKeyCoder } from './services/LeafKeyCoder';
+import { APIClient } from './services/APIClient';
+import { LeafValueCoder } from './services/LeafValueCoder';
+import { Registry as ContractRegistry } from './contracts/Registry';
+import { ChainContract } from './contracts/Chain';
+import { SortedMerkleTree } from './lib/SortedMerkleTree';
 import * as ABI from './contracts/abi';
 
 export {
-  LeafType,
   IAPIClientOptions,
   LeafValueCoder,
   APIClient,
   LeafKeyCoder,
   ContractRegistry,
   ChainContract,
-  converters,
   ABI,
+  SortedMerkleTree,
 };