+++ src/models/LeafType.ts
@@ -1,5 +1,0 @@
-export enum LeafType {
-  TYPE_HEX = 1,
-  TYPE_INTEGER,
-  TYPE_FLOAT
-}